coinbase/requirements.txt
coinbase-advanced-py
matplotlib
mypy
numpy
pyqt5
python-dateutil
scipy
sphinx
sphinx-rtd-theme
typeshed
